INT-8699: Redis fixes and improvements
There was a bug with the set_many method
in Redis cache.  Switched over to core's
testing method which replaced a lot of our
own custom testing.  This helped to identify
bugs and ensure we are doing what core expects.

In addition, added some code to allow Redis to
be used as the cache store during PHPUnit runs.

Note: there is an annoying bug that I had to
workaround in the Redis constructor.  On purge
all, the name and config were different from
the instance being used during testing.  So,
overwrite those the correct ones.